ci: Switch CI to qwen
It looks like upstream will also be going with qwen for
vllm integrations tests, so the tests are more likely to
be suited to it.

Also enable test_inference_store_tool_calls tests as it's
more capable at calling tools.

Disable test_text_chat_completion_non_streaming as the reasoning
can take too long and timeout. We can re-enable this when the
tests are adjusted upstream.

Closes: #77

Signed-off-by: Derek Higgins <derekh@redhat.com>